django/requirements/base.txt
Mezzanine==4.2.2
gunicorn==19.3.0
Django~=1.10.0
Pillow==2.8.1
boto==2.38.0
dj-database-url==0.3.0
django-dotenv~=1.4.1
django-extensions==1.7.6
django-grappelli==2.8.1
django-require==1.0.11
django-storages==1.1.8
django-model-utils==2.5
django-taggit==0.18.1
django-taggit-autosuggest==0.2.8
djangorestframework==3.1.3
django-admin-sortable2==0.6.8
mimeparse==0.1.3
celery==3.1.18
psycopg2==2.6
pytz==2015.4
pyuploadcare==2.0.1
django-wysiwyg-redactor==0.4.9.1
django-robots==2.0
pytest-django==3.1.2
pytest-pythonpath==0.7.1
